ExamVersionDAO feito

+++ exams/src/main/java/ras/exams/exams/data/ExamAnswerDAO.java
@@ -56,6 +56,70 @@ public class ExamAnswerDAO implements Map<UUID, ExamAnswer> {
         return ExamAnswerDAO.singleton;
     }
 
+    public List<ExamAnswer> getExamAnswersFromExam (UUID examID)
+    {
+        List<ExamAnswer> examAnswers = new ArrayList<>();
+        try (Connection conn = DriverManager.getConnection(DAOconfig.URL, DAOconfig.USERNAME, DAOconfig.PASSWORD);
+            Statement stm = conn.createStatement();
+            ResultSet rs = stm.executeQuery("""
+            SELECT BIN_TO_UUID(examAnswerID) as examAnswerID,
+                    BIN_TO_UUID(studentID) as studentID,
+                    grade
+            FROM examanswer
+            WHERE examID=UUID_TO_BIN('"""+examID.toString()+"')"))
+        {
+            if (rs.next())
+            {
+                UUID examAnswerID = UUID.fromString(rs.getString("examAnswerID")),
+                        studentID = UUID.fromString(rs.getString("studentID"));
+                int grade = rs.getInt("grade");
+                examAnswers.add(new ExamAnswer(examAnswerID, 
+                                    examID, 
+                                    studentID, 
+                                    grade, 
+                                    this.answerDAO.getAnswersFromExamAnswer(examAnswerID)));
+            }
+        }
+        catch (SQLException e)
+        {
+            e.printStackTrace();
+            throw new NullPointerException(e.getMessage());
+        }
+        return examAnswers; 
+    }
+
+    public ExamAnswer getStudentExamAnswer (UUID studentID)
+    {
+        ExamAnswer a = null;
+        try (Connection conn = DriverManager.getConnection(DAOconfig.URL, DAOconfig.USERNAME, DAOconfig.PASSWORD);
+            Statement stm = conn.createStatement();
+            ResultSet rs = stm.executeQuery("""
+            SELECT BIN_TO_UUID(examAnswerID) as examAnswerID,
+                    BIN_TO_UUID(examID) as examID,
+                    grade
+            FROM examanswer
+            WHERE studentID=UUID_TO_BIN('"""+studentID.toString()+"')"))
+        {
+            if (rs.next())
+            {
+                UUID examAnswerID = UUID.fromString(rs.getString("examAnswerID")),
+                        examID = UUID.fromString(rs.getString("examID"));
+                int grade = rs.getInt("grade");
+                a = new ExamAnswer(examAnswerID, 
+                                    examID, 
+                                    studentID, 
+                                    grade, 
+                                    this.answerDAO.getAnswersFromExamAnswer(examAnswerID));
+            }
+        }
+        catch (SQLException e)
+        {
+            e.printStackTrace();
+            throw new NullPointerException(e.getMessage());
+        }
+        return a;
+    }
+
     @Override
     public void clear() {
         try (Connection conn = DriverManager.getConnection(DAOconfig.URL, DAOconfig.USERNAME, DAOconfig.PASSWORD); Statement stm = conn.createStatement())
@@ -76,7 +140,7 @@ public class ExamAnswerDAO implements Map<UUID, ExamAnswer> {
         boolean r;
         try (Connection conn = DriverManager.getConnection(DAOconfig.URL, DAOconfig.USERNAME, DAOconfig.PASSWORD);
             Statement stm = conn.createStatement();
-            ResultSet rs = stm.executeQuery("SELECT examAnswerID FROM examanswer WHERE examAnswerID=UUID_TO_BIN('"+key.toString()+")'"))
+            ResultSet rs = stm.executeQuery("SELECT examAnswerID FROM examanswer WHERE examAnswerID=UUID_TO_BIN('"+key.toString()+"')"))
         {
             r = rs.next();
         }
@@ -175,7 +239,7 @@ public class ExamAnswerDAO implements Map<UUID, ExamAnswer> {
         Set<UUID> r = new HashSet<>();
         try (Connection conn = DriverManager.getConnection(DAOconfig.URL, DAOconfig.USERNAME, DAOconfig.PASSWORD);
             Statement stm = conn.createStatement();
-            ResultSet rs = stm.executeQuery("SELECT examAnswerID FROM examanswer"))
+            ResultSet rs = stm.executeQuery("SELECT BIN_TO_UUID(examAnswerID) FROM examanswer"))
         {
             while (rs.next())
             {
@@ -237,9 +301,10 @@ public class ExamAnswerDAO implements Map<UUID, ExamAnswer> {
             Statement stm = conn.createStatement())
         {
             stm.execute("SET FOREIGN_KEY_CHECKS=0");
-            stm.executeUpdate("DELETE FROM examAnswer WHERE examAnswerID=UUID_TO_BIN('"+key.toString()+"')");
-            for (Answer a : rv.getAnswers())
-                this.answerDAO.remove(a.getAnswerID());
+            stm.executeUpdate("DELETE FROM examanswer WHERE examAnswerID=UUID_TO_BIN('"+key.toString()+"')");
+            if (rv.getAnswers() != null)
+                for (Answer a : rv.getAnswers())
+                    this.answerDAO.remove(a.getAnswerID());
             stm.execute("SET FOREIGN_KEY_CHECKS=1");
         }
         catch (SQLException e)
